pyproject.toml
[tool.poetry]
name = "school_project"
version = "0.1.0"
description = ""
authors = ["Toren C <theplatecrafter@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
numpy-stl = "^3.1.1"
pytest = "^8.1.1"
matplotlib = "^3.8.4"
numpy = "^1.26.4"
ezdxf = "^1.2.0"
py3d = "^0.1.33"
scikit-image = "^0.23.2"
shapely = "^2.0.4"
moviepy = "^1.0.3"

[tool.poetry.dev-dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
